Add all weighted shortest path
