+++ src/include/function/gds/auxiliary_state/path_auxiliary_state.h
@@ -19,5 +19,24 @@ private:
     std::unique_ptr<BFSGraph> bfsGraph;
 };
 
+struct WSPPathsAuxiliaryState : public GDSAuxiliaryState {
+public:
+    explicit WSPPathsAuxiliaryState(std::unique_ptr<BFSGraph> bfsGraph)
+        : bfsGraph{std::move(bfsGraph)} {}
+
+    void initSource(common::nodeID_t sourceNodeID) override {
+        sourceParent.setCost(0);
+        bfsGraph->setParentList(sourceNodeID, &sourceParent);
+    }
+
+    void beginFrontierCompute(common::table_id_t, common::table_id_t toTableID) override {
+        bfsGraph->pinTableID(toTableID);
+    }
+
+private:
+    std::unique_ptr<BFSGraph> bfsGraph;
+    ParentList sourceParent;
+};
+
 } // namespace function
 } // namespace kuzu
