+++ src/planner/join_order/cardinality_estimator.cpp
@@ -240,7 +240,8 @@ double CardinalityEstimator::getExtensionRate(const RelExpression& rel,
     }
     case QueryRelType::SHORTEST:
     case QueryRelType::ALL_SHORTEST:
-    case QueryRelType::WEIGHTED_SHORTEST: {
+    case QueryRelType::WEIGHTED_SHORTEST:
+    case QueryRelType::ALL_WEIGHTED_SHORTEST: {
         auto rate = std::min<double>(
             oneHopExtensionRate * std::max<uint16_t>(rel.getUpperBound(), 1), numRels);
         return rate * context->getClientConfig()->recursivePatternCardinalityScaleFactor;
